ai-storyteller-deployable/requirements.txt
streamlit>=1.33
python-dotenv>=1.0.1
openai>=1.30.0
pydantic>=2.7.0
jinja2>=3.1.4
pyttsx3>=2.90
Pillow>=10.3.0
soundfile>=0.12.1
numpy>=1.26.4
